requirements.txt
colorama==0.4.6
iniconfig==2.0.0
numpy==2.0.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pycryptodome==3.20.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
pyzbar==0.1.9
six==1.16.0
tzdata==2024.1
